move aggregates out into their own file
