+++ src/runtime/runtime.ts
@@ -2686,25 +2686,6 @@ export abstract class AggregateNode extends Node {
 
 }
 
-type SumAggregateState = {total:number};
-export class SumAggregate extends AggregateNode {
-  add(state:SumAggregateState, resolved:RawValue[]):any {
-    state.total += resolved[0] as number;
-    return state;
-  }
-  remove(state:SumAggregateState, resolved:RawValue[]):any {
-    state.total -= resolved[0] as number;
-    return state;
-  }
-  getResult(state:SumAggregateState):RawValue {
-    return state.total;
-  }
-  newResultState():SumAggregateState {
-    return {total: 0};
-  };
-}
-
-
 //------------------------------------------------------------------------------
 // Block
 //------------------------------------------------------------------------------
